daemon: move wallet `jsonrpc_` methods to an individual module
